separate gradient tests and add more gradient tests.
