+++ src/node/tests.rs
@@ -179,136 +179,3 @@ fn test_multiple_computation() {
 
     assert_eq!(y.calculate().unwrap().get_scalar_f32(), Ok(-5.));
 }
-
-#[test]
-fn test_grad_self() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let x = Node::from_scalar(&hw, &g, 42.);
-
-    let gx = grad(x, &[x]).unwrap();
-    assert_eq!(gx.len(), 1);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(1.));
-}
-
-#[test]
-fn test_grad_unrelated() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let x = Node::from_scalar(&hw, &g, 42.);
-    let y = Node::from_scalar(&hw, &g, 42.);
-
-    let gx = grad(y, &[x]).unwrap();
-    assert_eq!(gx.len(), 1);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(0.));
-}
-
-#[test]
-fn test_grad_neg() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let x = Node::from_scalar(&hw, &g, 42.);
-    let y = -x;
-
-    let gx = grad(y, &[x]).unwrap();
-    assert_eq!(gx.len(), 1);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(-1.));
-}
-
-#[test]
-fn test_grad_add() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let a = Node::from_scalar(&hw, &g, 123.);
-    let b = Node::from_scalar(&hw, &g, 456.);
-    let y = a + b;
-
-    let gx = grad(y, &[a, b]).unwrap();
-    assert_eq!(gx.len(), 2);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(1.));
-
-    assert_eq!(gx[1].shape(), make_shape![]);
-    assert!(ptr::eq(gx[1].hardware(), &hw));
-    assert_eq!(gx[1].calculate().unwrap().get_scalar_f32(), Ok(1.));
-}
-
-#[test]
-fn test_grad_sub() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let a = Node::from_scalar(&hw, &g, 123.);
-    let b = Node::from_scalar(&hw, &g, 456.);
-    let y = a - b;
-
-    let gx = grad(y, &[a, b]).unwrap();
-    assert_eq!(gx.len(), 2);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(1.));
-
-    assert_eq!(gx[1].shape(), make_shape![]);
-    assert!(ptr::eq(gx[1].hardware(), &hw));
-    assert_eq!(gx[1].calculate().unwrap().get_scalar_f32(), Ok(-1.));
-}
-
-#[test]
-fn test_grad_mul() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let a = Node::from_scalar(&hw, &g, 123.);
-    let b = Node::from_scalar(&hw, &g, 456.);
-    let y = a * b;
-
-    let gx = grad(y, &[a, b]).unwrap();
-    assert_eq!(gx.len(), 2);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(456.));
-
-    assert_eq!(gx[1].shape(), make_shape![]);
-    assert!(ptr::eq(gx[1].hardware(), &hw));
-    assert_eq!(gx[1].calculate().unwrap().get_scalar_f32(), Ok(123.));
-}
-
-#[test]
-fn test_grad_div() {
-    let hw = RefCell::new(CpuHardware::new());
-    let g = RefCell::new(Graph::new());
-
-    let a = Node::from_scalar(&hw, &g, 3.);
-    let b = Node::from_scalar(&hw, &g, 2.);
-    let y = a / b;
-
-    let gx = grad(y, &[a, b]).unwrap();
-    assert_eq!(gx.len(), 2);
-
-    assert_eq!(gx[0].shape(), make_shape![]);
-    assert!(ptr::eq(gx[0].hardware(), &hw));
-    // dy/da == 1/b
-    assert_eq!(gx[0].calculate().unwrap().get_scalar_f32(), Ok(0.5));
-
-    assert_eq!(gx[1].shape(), make_shape![]);
-    assert!(ptr::eq(gx[1].hardware(), &hw));
-    // dy/db == -a/b^2
-    assert_eq!(gx[1].calculate().unwrap().get_scalar_f32(), Ok(-0.75));
-}
